Avoid downloading pdf menu if unchanged since last download, use bash syntax

+++ mdj.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 # Install dependencies
 TABULA=tabula-1.0.1-jar-with-dependencies.jar
@@ -15,26 +15,32 @@ cat menus.html | tidy -asxhtml -indent - 2> /dev/null | \
     sed s'/&nbsp;/ /g' > menus.xml
 
 # Extract link to the monthly menu PDF file
-#PDF_URL=`xpath menus.xml '//div[@class="box-download"]/a/@href' 2> /dev/null | \
-#    cut -c8- | sed s'/.$//'`
+#PDF_URL=$(xpath menus.xml '//div[@class="box-download"]/a/@href' 2> /dev/null | \
+    #    cut -c8- | sed s'/.$//')
 # the upper xpath expression doesn't seem to work with xmllint or xmlstarlet...
-PDF_URL=`cat menus.xml | grep -A2 download | \
-    tail -1 | tr -d ' ' | cut -c2- | sed s'/.$//'`
+PDF_URL=$(cat menus.xml | grep -A2 download | \
+    tail -1 | tr -d ' ' | cut -c2- | sed s'/.$//')
 PDF_URL="http://www.lfitokyo.org/$PDF_URL"
 
 # Download month menu PDF file
-wget -q $PDF_URL -O menu.pdf
+LAST_MOD_F=menu.pdf.last_modif_time
+LAST_MOD=$(curl -sI $PDF_URL | grep Last-Modified)
+if [ ! -f $LAST_MOD_F ] || [ "$LAST_MOD" != "$(cat $LAST_MOD_F)" ]; then
+  echo "Downloading file"
+  wget -q $PDF_URL -O menu.pdf
+fi
+echo $LAST_MOD > $LAST_MOD_F
 
 # Extract table from the 
 java -Dfile.encoding=utf-8 -jar $TABULA -gti -fCSV menu.pdf 2> /dev/null > menu.csv
 
 # Count columns
-COL_COUNT=`awk -F, '{print NF}' menu.csv | sort -nu | tail -n 1`
+COL_COUNT=$(awk -F, '{print NF}' menu.csv | sort -nu | tail -n 1)
 
 # Merge columns
 rm -f menu.txt
 touch menu.txt
-COLS=`seq $COL_COUNT`
+COLS=$(seq $COL_COUNT)
 for i in $COLS; do
   cut -d, -f $i menu.csv >> menu.txt
 done
@@ -49,13 +55,13 @@ sed -i s'/  */ /g' menu.txt
 echo "" >> menu.txt
 
 # Compute current date in French
-TODAY=`LC_TIME='fr_FR.UTF-8' date '+%A %d'`
-TODAY_UC=`echo $TODAY | awk '{print toupper($0)}'`
+TODAY=$(LC_TIME='fr_FR.UTF-8' date '+%A %d')
+TODAY_UC=$(echo $TODAY | awk '{print toupper($0)}')
 
 # Output today's menu
-MENU=`cat menu.txt | \
+MENU=$(cat menu.txt | \
     sed -n -e "/$TODAY_UC/,/LUNDI\|MARDI\|MERCREDI\|JEUDI\|VENDREDI/p" | \
-    tail -n +2 | head -n -1`
+    tail -n +2 | head -n -1)
 
 curl -s -F "token=$PUSHOVER_KEY" \
     -F "user=$PUSHOVER_USER" \
